chore: organise workspaces by domain
