+++ auth/feature-tests/vitest-step-definitions/GovUKMobileCognitoWAFAssociation.steps.ts
@@ -12,7 +12,7 @@ let template: Template;
 
 describeFeature(feature, ({ BeforeAllScenarios, Scenario }) => {
   BeforeAllScenarios(() => {
-    let yamltemplate = load(readFileSync("template.yaml", "utf-8"), {
+    const yamltemplate = load(readFileSync("template.yaml", "utf-8"), {
       schema: schema,
     });
     template = Template.fromJSON(yamltemplate);
