+++ auth/feature-tests/vitest-step-definitions/GovUKMobileResourceTags.steps.ts
@@ -20,7 +20,7 @@ const ignoredResources = [
 
 describeFeature(feature, ({ BeforeAllScenarios, Scenario }) => {
   BeforeAllScenarios(() => {
-    let yamltemplate = load(readFileSync("template.yaml", "utf-8"), {
+    const yamltemplate = load(readFileSync("template.yaml", "utf-8"), {
       schema: schema,
     });
     template = Template.fromJSON(yamltemplate);
