+++ auth/hello-world/.eslintignore
