+++ auth/hello-world/.eslintrc.js
