+++ auth/hello-world/.npmignore
