+++ auth/hello-world/.prettierrc.js
