+++ auth/hello-world/package.json
