+++ auth/hello-world/tsconfig.json
