+++ auth/package.json
@@ -1,6 +1,6 @@
 {
     "name": "auth",
-    "description": "GOVUK Mobile Auth",
+    "description": "Auth",
     "version": "0.0.1",
     "private": true,
     "devDependencies": {
@@ -14,6 +14,7 @@
         "yaml-cfn": "^0.3.2"
     },
     "scripts": {
+        "test:infrastructure": "vitest run",
         "test:validate": "sam validate",
         "test": "vitest"
     }
