+++ auth/template.yaml
@@ -52,6 +52,8 @@ Globals:
       - UsePermissionsBoundary
       - !Ref PermissionsBoundary
       - !Ref AWS::NoValue
+    MemorySize: 128
+    Timeout: 100
 
 Resources:
   GovUKMobileCognitoUserPool:
@@ -143,9 +145,8 @@ Resources:
   GovUKMobileTestFunction:
     Type: AWS::Serverless::Function
     Properties:
-      InlineCode: |
-        exports.handler = async (event) => ++event;
-      Handler: app.lambda_handler
+      CodeUri: src
+      Handler: src/handlers/hello-world.handler
       Runtime: nodejs22.x
       Events:
         HelloWorldApi:
@@ -155,6 +156,13 @@ Resources:
               Ref: GovUKMobileApiGateway
             Path: /test
             Method: GET
+      Tags:
+        - Key: Product
+          Value: GOV.UK
+        - Key: Environment
+          Value: !Ref Environment
+        - Key: System
+          Value: Authentication
 
 Outputs:
   GovUKMobileApiGateway:
