+++ package.json
@@ -20,7 +20,7 @@
     "typescript-eslint": "^8.29.1"
   },
   "workspaces": [
-    "sam-app/*"
+    "auth/*"
   ],
   "optionalDependencies": {
     "@rollup/rollup-linux-x64-gnu": "4.9.5"
