--- sam-app/feature-tests/vitest-features/GovUKMobileApiGateway.feature
@@ -1,4 +1,0 @@
-Feature: GOV UK Api Gateway
-    Scenario: A template can deploy the GOV UK Api Gateway
-        Given a template to deploy the GOV UK Api Gateway
-        Then the template must have the required resource and properties to deploy the GOV UK Api Gateway
