--- sam-app/feature-tests/vitest-features/GovUKMobileCognitoUserPool.feature
@@ -1,4 +1,0 @@
-Feature: GOV UK Mobile Cognito UserPool
-    Scenario: A template can deploy the GOV UK Mobile Cognito Userpool 
-        Given a template to deploy the GOV UK Mobile Cognito Userpool
-        Then the template must have the required resource and properties to deploy the GOV UK Mobile Cognito Userpool
