--- sam-app/feature-tests/vitest-features/GovUKMobileResourceTags.feature
@@ -1,4 +1,0 @@
-Feature: Resources are tagged correctly
-    Scenario: A template has the correct resource tags 
-        Given a template with AWS resources 
-        Then the template's resources must have the required tags
