--- sam-app/feature-tests/vitest-features/GovUKMobileWebApplicationFirewall.feature
@@ -1,4 +1,0 @@
-Feature: GOV UK Mobile Web Application Firewall 
-    Scenario: A template can deploy the GOV UK Mobile Web Application Firewall 
-        Given a template to deploy the GOV UK Mobile Web Application Firewall
-        Then the template must have the required resource and properties to deploy the GOV UK Mobile Web Application Firewall
