--- sam-app/template.yaml
@@ -1,170 +1,0 @@
-AWSTemplateFormatVersion: 2010-09-09
-Transform: AWS::Serverless-2016-10-31
-Description: >
-  devplatform-deploy canary-circuit-breaker template version: v0.0.0
-  Canary deployment circuit breaker template managed by dev-platform.
-
-Parameters:
-  Environment:
-    Description: "The name of the environment to deploy to"
-    Type: "String"
-    AllowedValues:
-      - build
-      - staging
-      - production
-      - integration
-      - demo
-      - local
-      - dev
-
-  CodeSigningConfigArn:
-    Type: String
-    Description: >
-      The ARN of the Code Signing Config to use, provided by the deployment pipeline
-    Default: "none"
-    
-  PermissionsBoundary:
-    Type: String
-    Description: >
-      The ARN of the permissions boundary to apply to any role created by the template
-    Default: "none"
-    
-Conditions:
-  UseCodeSigning:
-    Fn::Not:
-      - Fn::Equals:
-          - !Ref CodeSigningConfigArn
-          - "none"
-
-  UsePermissionsBoundary:
-    Fn::Not:
-      - Fn::Equals:
-          - !Ref PermissionsBoundary
-          - "none"
-
-Globals:
-  Function:
-    CodeSigningConfigArn: !If
-      - UseCodeSigning
-      - !Ref CodeSigningConfigArn
-      - !Ref AWS::NoValue
-    PermissionsBoundary: !If
-      - UsePermissionsBoundary
-      - !Ref PermissionsBoundary
-      - !Ref AWS::NoValue
-    MemorySize: 128
-    Timeout: 100
-
-Resources:
-  GovUKMobileCognitoUserPool:
-    Type: AWS::Cognito::UserPool
-    Properties:
-      UserPoolName: !Join
-        - "-"
-        - - !Ref AWS::StackName
-          - "user-pool"
-          - Fn::Select:
-              - 4
-              - Fn::Split:
-                  - "-"
-                  - Fn::Select:
-                      - 2
-                      - Fn::Split:
-                          - "/"
-                          - Ref: AWS::StackId
-      DeletionProtection: ACTIVE
-      UsernameAttributes:
-        - email
-      Schema:
-        - AttributeDataType: String
-          Name: email
-          Required: true
-      UserPoolTags:
-        Product: GOV.UK
-        Environment: !Ref Environment
-        System: Authentication
-
-  GovUKMobileWebApplicationFirewall:
-    Type: AWS::WAFv2::WebACL
-    Properties:
-      DefaultAction:
-        Allow: {}
-      Scope: REGIONAL
-      VisibilityConfig:
-        CloudWatchMetricsEnabled: true
-        MetricName: !Join
-          - "-"
-          - - !Ref AWS::StackName
-            - "waf-acl-rules"
-            - Fn::Select:
-                - 4
-                - Fn::Split:
-                    - "-"
-                    - Fn::Select:
-                        - 2
-                        - Fn::Split:
-                            - "/"
-                            - Ref: AWS::StackId
-        SampledRequestsEnabled: true
-      Tags:
-        - Key: Product
-          Value: GOV.UK
-        - Key: Environment
-          Value: !Ref Environment
-        - Key: System
-          Value: Authentication
-
-  GovUKMobileCognitoWAFAssociation:
-    Type: AWS::WAFv2::WebACLAssociation
-    Properties:
-      ResourceArn: !GetAtt GovUKMobileCognitoUserPool.Arn
-      WebACLArn: !GetAtt GovUKMobileWebApplicationFirewall.Arn
-
-  GovUKMobileApiGateway:
-    Type: AWS::Serverless::Api
-    Properties:
-      Name: !Join
-        - "-"
-        - - !Ref AWS::StackName
-          - "api-gateway"
-          - Fn::Select:
-              - 4
-              - Fn::Split:
-                  - "-"
-                  - Fn::Select:
-                      - 2
-                      - Fn::Split:
-                          - "/"
-                          - Ref: AWS::StackId
-      StageName: !Ref Environment
-      Tags:
-        Product: GOV.UK
-        Environment: !Ref Environment
-        System: Authentication
-
-  GovUKMobileTestFunction:
-    Type: AWS::Serverless::Function
-    Properties:
-      CodeUri: src
-      Handler: src/handlers/hello-world.handler
-      Runtime: nodejs22.x
-      Events:
-        HelloWorldApi:
-          Type: Api
-          Properties:
-            RestApiId:
-              Ref: GovUKMobileApiGateway
-            Path: /test
-            Method: GET
-      Tags:
-        - Key: Product
-          Value: GOV.UK
-        - Key: Environment
-          Value: !Ref Environment
-        - Key: System
-          Value: Authentication
-
-Outputs:
-  GovUKMobileApiGateway:
-    Description: API Gateway endpoint URL Hello World function
-    Value: !Sub https://${GovUKMobileApiGateway}.execute-api.${AWS::Region}.amazonaws.com/${Environment}/test/
