--- sam-app/tsconfig.json
@@ -1,20 +1,0 @@
-{
-    "compilerOptions": {
-        "target": "es2020",
-        "strict": true,
-        "preserveConstEnums": true,
-        "noEmit": true,
-        "sourceMap": false,
-        "module": "esnext",
-        "moduleResolution": "node",
-        "resolveJsonModule": true,
-        "esModuleInterop": true,
-        "skipLibCheck": true,
-        "forceConsistentCasingInFileNames": true,
-        "baseUrl": "./"
-    },
-    "exclude": [
-        "node_modules",
-        "**/*.test.ts"
-    ],
-}
