--- sam-app/vitest.config.ts
@@ -1,10 +1,0 @@
-import { defineConfig } from "vitest/config"
-import { VitestCucumberPlugin } from "@amiceli/vitest-cucumber"
-
-export default defineConfig({
-    test: {
-        include: [
-            "**/*.steps.ts",
-        ]
-    }
-})
