+++ tsconfig.json
@@ -4,7 +4,7 @@
   "files": [],
   "references": [
     {
-      "path": "./sam-app"
+      "path": "./auth"
     },
   ]
 }
